+++ src/env/web/process.ts
@@ -1,0 +1,5 @@
+import { ThisExtension } from "../../ThisExtension";
+
+export async function startExternalProcess(executable: string, args: string[]): Promise<void> {
+	ThisExtension.log(`Starting external process is not supported in the web!`)
+}